Alertas, edición y eliminación correctamente

+++ src/screens/SingleSavingScreen.jsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import { ScrollView, StyleSheet, View } from 'react-native';
+import { ScrollView, StyleSheet, View, Alert } from 'react-native';
 import { useRoute } from '@react-navigation/native';
 import SingleSavingCard from '../components/savings/SingleSavingCard';
 import FAB from '../components/FAB';
@@ -20,21 +19,33 @@ const SingleSavingScreen = () => {
     const { saving } = route.params;
 
     const handleDelete = async ()=>{
-        try{
-            await deleteSaving(saving.id)
-            navigation.goBack()
-            console.log("Ahorro eliminado:", saving.name)
-        }catch{
-            console.log("Error al eliminar ahorro")
-        }
+        Alert.alert(
+            'Eliminar Ahorro',
+            `¿Estás seguro de que quieres eliminar el ahorro "${saving.name}"? Esta acción no se puede deshacer.`,
+            [
+                { text: 'cancelar', style: 'cancel'},
+                { text: 'Eliminar', syle: 'destructive', onPress: async ()=> {
+                    
+                    try{
+                        await deleteSaving(saving.id)
+                        navigation.goBack()
+                        console.log("Ahorro eliminado:", saving.name)
+                    }catch{
+                        console.log("Error al eliminar ahorro")
+                    }
+
+                }}
+            ]
+        )
     }
+
     
     return (
         <View style={{ flex: 1, backgroundColor: colors.background }}>
         <ScrollView contentContainerStyle={styles.content}>
         <SingleSavingCard
             saving={saving}
-            onEdit={() => console.log('Editar ahorro')}
+            onEdit={() => navigation.navigate('AddSavingScreen', { toEdit: saving })}
             onDelete={handleDelete}
         />
         <View style={{ paddingVertical: 24 }}>
